Crypto currency price prediction using machine learning techniques
Predicting the price of Crypto currencies like Bitcoin, Dogecoin.. etc is tough because it does not have any proper   fundamentals so it’s very risky to trade them in the Crypto market. Since Crypto predictions are unreliable, we need a proper way to predict them to avoid loss while trading in the crypto market. So, the main objective of our project    is to implement machine learning techniques to predict the price of crypto currencies in the market.
  Tools Used: NumPy, pandas, matplotlib, Pandas data reader, tenser flow, scikit-learn